+++ react-app/src/modules/user/update-pages/components/sideBar/left-side-bar.css
@@ -1,0 +1,12 @@
+.update_page_background {
+    background-image: linear-gradient(226deg, rgb(253 209 49) 48%, rgb(255 250 88) 100%);
+    padding-top: 90px;
+    height: 100vh;
+    border-right: 1px solid #cdc64e;
+}
+
+.page_return {
+    margin: 15px 0 0 20px;
+    cursor: pointer;
+    width: 30px;
+}
