+++ react-app/src/modules/user/update-pages/components/updatePage/updatePage.css
@@ -1,0 +1,16 @@
+.update_page_content {
+    padding-top: 90px;
+}
+.update_page_form {
+    margin: 50px 30px
+}
+.update_page_form_title {
+    font-size: 30px;
+    font-weight: 350;
+}
+.update_title_text {
+    margin-right: 7px;
+}
+.update_title_body {
+    margin-top: 20px;
+}
