+++ react-app/src/modules/user/update-pages/pages/update-user-page/components/update.from.css
@@ -1,29 +1,3 @@
-.update_page_background {
-    background-image: linear-gradient(226deg, rgb(253 209 49) 48%, rgb(255 250 88) 100%);
-    padding-top: 90px;
-    height: 100vh;
-    border-right: 1px solid #cdc64e;
-}
-.container {
-    padding: 0;
-    margin: 0;
-}
-.update_page_content {
-    padding-top: 90px;
-}
-.update_page_form {
-    margin: 50px 30px
-}
-.update_page_form_title {
-    font-size: 30px;
-    font-weight: 350;
-}
-.update_title_text {
-    margin-right: 7px;
-}
-.update_title_body {
-    margin-top: 20px;
-}
 .update_form {
     width: 300px;
 }
