+++ react-app/src/modules/user/update-pages/pages/update-user-page/update-user-page.css
@@ -1,0 +1,4 @@
+.container {
+    padding: 0;
+    margin: 0;
+}
