+++ react-app/src/modules/user/update-pages/pages/update-user-page/updateUserPage.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import '../../../../../core/css/styles.css';
+import './update-user-page.css';
+import { ToolIcon } from '../../../../../core/components/icons';
+import { LeftSideBar } from '../../components/sideBar/left-side-bar';
+import { UpdatePage } from '../../components/updatePage/updatePageTempate';
+import { UpdateUserFrom } from './components/update.form';
+
+const UpdateUserPage: React.FC = () => {
+  return (
+    <div className="app">
+      <div className="container">
+        <div className="row">
+          <LeftSideBar/>
+          <UpdatePage header={ 'Update User' } Icon={ <ToolIcon/> }>
+            <UpdateUserFrom/>
+          </UpdatePage>
+        </div>
+      </div>
+    </div>
+  );
+};
+export default UpdateUserPage
