+++ react-app/src/pages/user-page/update-user-pages/update-user-page/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import UpdateUserPage from '../../../../modules/user/update-pages/update-user-page/updateUserPage';
+import UpdateUserPage from '../../../../modules/user/update-pages/pages/update-user-page/updateUserPage';
 
 const UpdateUser = () => {
